% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMwind.R
\docType{data}
\name{QMwind}
\alias{QMwind}
\title{QMwind Data}
\format{
An object of class \code{griddap_nc} (inherits from \code{nc}, \code{data.frame}) with 0 rows and 2 columns.
}
\usage{
QMwind
}
\description{
pre-Download of QMwind in `add_griddap()` example so that example
can run within CRAN Time limits
}
\details{
obtained using the `rerddap` command
wind <- griddap('erdQMwindmday',
       time = c('2016-11-16', '2017-01-16'), latitude = c(30, 50),
       longitude = c(210, 240),
       fields = 'x_wind')
)
}
\keyword{datasets}
