% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_set.R
\name{bbox_set}
\alias{bbox_set}
\title{change bounding box in plotdap object}
\usage{
bbox_set(plotobj, xlim, ylim)
}
\arguments{
\item{plotobj}{valid plotdap object}

\item{xlim}{new x-values of the bounding box}

\item{ylim}{new y-values of the bounding box}
}
\value{
a plotdap object
}
\description{
\code{bbox_set}changes the bounding box in an plotdap object.
Particularly needed if using gganimate::animate()
}
\examples{
p <- plotdap()
p <- add_tabledap(p, sardines, ~subsample_count)
xlim = c(-125, -115)
ylim <- c(30., 50.)
p <- bbox_set(p, xlim, ylim)
}
